refactor(services): lift ssh session seam out of backup
The SshSession trait and MockSshSession are transport, not backup
logic. bichon rescan needs the same seam, and importing it from
services::backup::ssh would misstate the dependency.
