+++ src/services/backup.rs
@@ -2,5 +2,4 @@ pub mod executor;
 pub mod recipe;
 pub mod restic;
 pub mod session;
-pub mod ssh;
 pub mod verify;
